+++ srcs/get_va_uns.c
@@ -6,7 +6,7 @@
 /*   By: jboer <jboer@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/09/11 18:22:57 by jboer          #+#    #+#                */
-/*   Updated: 2019/09/26 14:24:36 by mvan-eng      ########   odam.nl         */
+/*   Updated: 2019/09/27 16:45:27 by mvan-eng      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,13 +68,13 @@ static void		ui_to_str(unsigned long long n, t_print *print)
 void			get_va_uns(t_print *print, va_list ap)
 {
 	if (print->spec == 0)
-		ui_to_str((unsigned long long)va_arg(ap, unsigned int), print);
+		ui_to_str(va_arg(ap, unsigned int), print);
 	if (print->spec == 1)
 		ui_to_str((unsigned char)va_arg(ap, unsigned int), print);
 	if (print->spec == 2)
 		ui_to_str((unsigned short)va_arg(ap, unsigned int), print);
 	if (print->spec == 3)
-		ui_to_str((unsigned long)va_arg(ap, unsigned long int), print);
+		ui_to_str(va_arg(ap, unsigned long), print);
 	if (print->spec == 4)
-		ui_to_str((unsigned long long)va_arg(ap, unsigned long long), print);
+		ui_to_str(va_arg(ap, unsigned long long), print);
 }
